Resolved crash on window downsizing
The downsizing of the window caused line 58 to throw an access violation exception.

[WHY?]
The frame instanly received the resizing upon the window size change event, while web-view surface resizing would happen later, once the mouse was released (this is described as "asynchronous resizing" somewhere in the source code). 
While upsizing the window nothing bad occurred, since the frame was bigger than the web-view surface, so the CopyTo command happened fine.
While downsizing the window an exception was thrown, since the frame was smaller than the web-view surface, so the CopyTo command would overflow the frame buffer.